Tarea #3004 - Eliminado el login de contacto y dejado solo para el usuario
