+++ Core/Table/contactos.xml
@@ -90,20 +90,8 @@
         <type>character varying(10)</type>
     </column>
     <column>
-        <name>lastactivity</name>
-        <type>timestamp</type>
-    </column>
-    <column>
-        <name>lastip</name>
-        <type>character varying(40)</type>
-    </column>
-    <column>
         <name>level</name>
         <type>integer</type>
-    </column>
-    <column>
-        <name>logkey</name>
-        <type>character varying(100)</type>
     </column>
     <column>
         <name>nombre</name>
@@ -112,10 +100,6 @@
     <column>
         <name>observaciones</name>
         <type>text</type>
-    </column>
-    <column>
-        <name>password</name>
-        <type>character varying(255)</type>
     </column>
     <column>
         <name>personafisica</name>
